Add files via upload
I've found the maximum quantum violation bound for CHSH inequality for any dimension of the operator, It's based on the simple yet elegant iterative algorithm which I've stated in the readme text file. I've used a RandomUnitary function from QetLab library, Rest of the code is written from scratch using Matlab. This code is subjective to change.

Thank you